// src/app/model/student.ts
import { ValidationMessage } from './validationmessage';

export class Student {
    "StudentID": String;
    "InstituteID": String;
    "BoardID": String;
    "MediumID": String;
    "UserID": String;
    "Rolename": String;
    "Name": String;
    "MiddleName": String;
    "LastName":String;
    "MotherName": String;
    "Gender": String;
    "DOB": String;
    "ClassID": String;
    "EASectionListID": String[];
    "SectionName": String;
    "AdmissionNumber": String;
    "Email": String;
    "CurrentRollNumber": String;
    "Address":String;
    "StudentCountryID": String;
    "StateID": String;
    "City": String;
    "Pincode": String;
    "AadharNumber": String;
    "Mobile": String;
    "SecondaryEmail": String;
    "CurrentAcademicSessionID": String;
    "PartnerCode": String;
    "ListUserValidationInfoMember": ValidationMessage[];
    "CountryName": String;
    "StateName": String;
    "objParentInfoMember": Object;
    "Status": Number;
    AESectionID: any;
    "AcademicYearID": String
}